requirements.txt
beautifulsoup4==4.11.1
pandas==1.5.2
Pillow==9.2.0
Requests==2.31.0
selenium==4.11.2
streamlit==1.28.0
tqdm==4.64.0
openpyxl==3.0.10
st-pages